cambios en la logica de consulta y pre consulta
